New stuff
Created a views folder so we can have different views files but for the same app

Edited the database a bit but you'll likely have to delete your version and migrate mine.

Added some paths to the urls as well. Use the below link to see the result of the query from the db I did (make sure your server is working is active first);
http://127.0.0.1:8000/projects/project_details/ProjectA/